==> notebooks/data retrieval/red link-candidate pairs/02. Get en id-name.ipynb <==
import codecs
import csv
import time
import os
import re
import gzip
import pandas as pd
import numpy as np

PATH_TO_DATA = '/media/andrii/earth/Katia/CS_MasterThesis/data/enwiki/en_parsed/'
data_time = '20180920'

ENWIKI_ART_FNMS = []
for file in os.listdir(PATH_TO_DATA):
    if file.endswith("art.csv.gz"):
        ENWIKI_ART_FNMS.append(file)
ENWIKI_RED_FNMS = []
for file in os.listdir(PATH_TO_DATA):
    if file.endswith("red.csv.gz"):
        ENWIKI_RED_FNMS.append(file)     

len(ENWIKI_ART_FNMS)

def unpack(file_name):
    file_name_new = file_name.replace(".gz","")
    with gzip.open(file_name, 'rb') as f_in, open(file_name_new, 'wb') as f_out:
        f_out.writelines(f_in)
    return file_name_new

def pack_and_remove(file_name):
    file_name_new = file_name+'.gz'
    with open(file_name, 'rb') as f_in, gzip.open(file_name_new, 'wb') as f_out:
        f_out.writelines(f_in)
    os.remove(file_name)
    return file_name_new

df_en_id_name = None
for fn in ENWIKI_ART_FNMS:
    fn = PATH_TO_DATA+fn
    print(fn)
    fn_new = unpack(fn)
    df_articles = pd.read_csv(fn_new, encoding='ISO-8859-1', quotechar="\"")
    df_en_id_name_tmp = df_articles[['id', 'title', 'text_len']].drop_duplicates()
    df_en_id_name_tmp.columns = ['id', 'title', 'length']
    #print("df_id_name_tmp size: {}".format(df_id_name_tmp.shape))
    if df_en_id_name is not None:
        df_en_id_name = df_en_id_name.append(df_en_id_name_tmp)        
        #print("append")
    else:
        df_en_id_name = df_en_id_name_tmp
        #print("assign")
    print("df_en_id_name size: {}".format(df_en_id_name.shape))
    os.remove(fn_new)

df_en_id_name = df_en_id_name.drop_duplicates()

df_en_id_name = df_en_id_name[['id', 'title']]
df_en_id_name.head()

f = lambda x: str(x).encode('ISO-8859-1')
df_en_id_name = df_en_id_name.applymap(f)

f_decode = lambda x: x.decode('utf8')
df_en_id_name = df_en_id_name.applymap(f_decode)

df_en_id_name.to_csv(PATH_TO_DATA+'enwiki-' + data_time + '-id_name.csv', index = False, encoding='utf8', sep ='^')

df_en_id_name_read = pd.read_csv(PATH_TO_DATA+'enwiki-' + data_time + '-id_name.csv', encoding='utf8', sep ='^')

df_en_id_name_read.head()
