==> notebooks/data retrieval/red link-candidate pairs/03. Get list of non-translated en articles.ipynb <==
import pandas as pd

PATH_TO_DATA_UK = '/media/andrii/earth/Katia/CS_MasterThesis/data/ukwiki/'
PATH_TO_DATA_EN = '/media/andrii/earth/Katia/CS_MasterThesis/data/enwiki/en_parsed/'
data_time = '20180920'

ukredlinks_with_en = pd.read_csv(PATH_TO_DATA_UK+'df_allukredlinks_withentitle.csv')

ukredlinks_with_en.head()

# add en ids for parent articles
en_idname = pd.read_csv(PATH_TO_DATA_EN+'enwiki-20180920-id_name.csv', sep='^')

en_idname.head()

ukredlinks_with_enids = pd.merge(ukredlinks_with_en, en_idname, how='left', left_on=['ll_title'], right_on=['title'])

ukredlinks_with_enids = ukredlinks_with_enids[['id_parent', 'red_link_name', 'red_link_id', 'id', 'title']]

ukredlinks_with_enids.head()

# read cleaned sample
ukredlinks_sample = pd.read_csv(PATH_TO_DATA_UK+'ukredlinks_sample.csv')

ukredlinks_with_enids_sample = pd.merge(ukredlinks_sample, ukredlinks_with_enids,
                                        how='left', left_on=['red_link_name'], right_on=['red_link_name'])

ukredlinks_with_enids_sample = ukredlinks_with_enids_sample[pd.notnull(ukredlinks_with_enids_sample['id'])]

ukredlinks_with_enids_sample = ukredlinks_with_enids_sample.drop_duplicates()

ukredlinks_with_enids_sample.to_csv(PATH_TO_DATA+"ukredlinks_with_enids_sample.csv", 
                     index = False, encoding='UTF-8', quotechar="\"", sep ="^")

ukredlinks_with_enids_sample.head()

# get non-translated English articles
fn = PATH_TO_DATA_UK+'20180920-langlinks_uk_en.csv'
df_link = pd.read_csv(fn,  encoding='UTF-8', quotechar='\'')

translated_en_articles = pd.merge(en_idname, df_link,
                                        how='left', left_on=['title'], right_on=['ll_title'])

translated_en_articles = translated_en_articles[['id', 'title', 'll_from']]

translated_en_articles = translated_en_articles[pd.notnull(translated_en_articles['ll_from'])]

non_translated_en_articles = translated_en_articles[translated_en_articles['ll_from'].isnull()]

translated_en_articles.to_csv(PATH_TO_DATA_EN+"translated_en_articles.csv", 
                     index = False, encoding='UTF-8', quotechar="\"", sep ="^")

non_translated_en_articles = non_translated_en_articles[['id', 'title']]

non_translated_en_articles.to_csv(PATH_TO_DATA_EN+"non_translated_en_articles.csv", 
                     index = False, encoding='UTF-8', quotechar="\"", sep ="^")
