==> notebooks/data retrieval/red link-candidate pairs/04. additional_ukredlink_data-preproc_check more.ipynb <==
import csv
import os
import gzip
import pandas as pd
import numpy as np

#path to data
PATH_TO_DATA = '/media/andrii/earth/Katia/CS_MasterThesis/data/'
PATH_TO_DATA_UK = PATH_TO_DATA+"ukwiki/"
PATH_TO_DATA_EN = PATH_TO_DATA+"enwiki/en_parsed/"

data_time = '20180920'

def unpack(file_name):
    file_name_new = file_name.replace(".gz","")
    with gzip.open(file_name, 'rb') as f_in, open(file_name_new, 'wb') as f_out:
        f_out.writelines(f_in)
    return file_name_new

def pack_and_remove(file_name):
    file_name_new = file_name+'.gz'
    with open(file_name, 'rb') as f_in, gzip.open(file_name_new, 'wb') as f_out:
        f_out.writelines(f_in)
    os.remove(file_name)
    return file_name_new

UKWIKI_ART_FNMS = []
for file in os.listdir(PATH_TO_DATA_UK):
    if file.endswith("art.csv"):
        UKWIKI_ART_FNMS.append(file)
UKWIKI_RED_FNMS = []
for file in os.listdir(PATH_TO_DATA_UK):
    if file.endswith("red.csv"):
        UKWIKI_RED_FNMS.append(file)   

print('UKWIKI_ART_FNMS:', UKWIKI_ART_FNMS)
print('UKWIKI_RED_FNMS:', UKWIKI_RED_FNMS)

df_uk_id_name = pd.read_csv(PATH_TO_DATA_UK+"ukwiki-" + data_time + "-id_name.csv", encoding='utf-8', sep="^")
df_uk_id_name = df_uk_id_name[['id', 'title']]
df_uk_id_name.head()

fn = PATH_TO_DATA_UK+UKWIKI_ART_FNMS[0]
print(fn)
df_articles = pd.read_csv(fn, encoding='UTF-8', quotechar="\"")
df_articles['link_val'] = df_articles['link_val'].str.split("#").str[0] # ???
df_articles = df_articles[df_articles['link_val'].str.strip() != '']
df_articles['link_pos_perc'] = (df_articles['link_pos']/df_articles['text_len']).round(3)    
df_articles['link_val_lc'] = df_articles['link_val'].str.lower()
df_articles.head(100)

df_articles[df_articles['link_val_lc'] == 'премія «золотий глобус» за найкращий  фільм — драма']

df_uk_titles_series = df_articles.iloc[:,7]

df_uk_titles_series_stripped = df_uk_titles_series.str.replace('  ', ' ')

df_titles_stripped = df_uk_titles_series_stripped.to_frame()

#replace not stripped in df_articles with df_titles_stripped
df_articles_less = df_articles[['id', 'title', 'link_val']]

df_articles = pd.concat([df_articles_less, df_titles_stripped], axis=1)

df_articles['link_val_lc'] = df_articles['link_val_lc'].str.strip()

df_articles = pd.merge(df_articles, df_uk_id_name_all, how='left', left_on=['link_val_lc'], right_on=['title_lc'])

df_articles.head()

df_articles = df_articles[['id_x', 'id_y', 'link_val']]
df_articles.columns = ['id', 'link_id', 'link_val']
df_articles.head()

df_articles = df_articles[pd.notnull(df_articles['link_val'])]

df_articles[['link_id']] = df_articles[['link_id']].fillna(-1).astype(int)
df_articles['is_red_link'] = np.where(df_articles['link_id'] == -1, True, False)

df_articles.to_csv(PATH_TO_DATA_UK+'ukwiki-20180920-pages-links.csv', index = False, encoding='UTF-8', quotechar="\"")

df_id_name = pd.read_csv(PATH_TO_DATA_EN+'enwiki-' + data_time + '-id_name.csv', sep="^")

df_id_name.head()

fn = PATH_TO_DATA_UK+"ukwiki-" + data_time + "-langlinks_en.csv"
df_en_langlinks = pd.read_csv(fn, encoding='UTF-16', sep="^")

df_en_langlinks = df_en_langlinks[pd.notnull(df_en_langlinks['ll_title'])]

df_en_langlinks.head()

#df_uk_en_links = pd.merge(df_en_langlinks, df_id_name, how='left', left_on=['ll_title'], right_on=['title'])

# ??????

df_uk_en_links = df_uk_en_links[['ll_from', 'id']]
df_uk_en_links.columns = ['id_uk', 'id_en']
df_uk_en_links[['id_uk']] = df_uk_en_links[['id_uk']].fillna(-1).astype(int)
df_uk_en_links[['id_en']] = df_uk_en_links[['id_en']].fillna(-1).astype(int)
df_uk_en_links.to_csv(PATH_TO_DATA_UK + data_time + "-langlinks_uk_en.csv", index = False, sep="^")

df_en_langlinks.to_csv(PATH_TO_DATA + data_time + "-langlinks_uk_en.csv", index = False, sep="^")
