==> notebooks/data retrieval/red link-candidate pairs/05. data-preproc-en_dump.ipynb <==
import csv
import os
import re
import gzip
import pandas as pd
import numpy as np

PATH_TO_DATA = '/media/andrii/earth/Katia/CS_MasterThesis/data/'
PATH_TO_DATA_EN = PATH_TO_DATA+"enwiki/en_parsed/"
PATH_TO_DATA_UK = PATH_TO_DATA+"ukwiki/"
data_time = '20180920'

def unpack(file_name):
    file_name_new = file_name.replace(".gz","")
    with gzip.open(file_name, 'rb') as f_in, open(file_name_new, 'wb') as f_out:
        f_out.writelines(f_in)
    return file_name_new

def pack_and_remove(file_name):
    file_name_new = file_name+'.gz'
    with open(file_name, 'rb') as f_in, gzip.open(file_name_new, 'wb') as f_out:
        f_out.writelines(f_in)
    os.remove(file_name)
    return file_name_new

ENWIKI_ART_FNMS = []
for file in os.listdir(PATH_TO_DATA_EN):
    if file.endswith("art.csv.gz"):
        ENWIKI_ART_FNMS.append(file)
ENWIKI_RED_FNMS = []
for file in os.listdir(PATH_TO_DATA_EN):
    if file.endswith("red.csv.gz"):
        ENWIKI_RED_FNMS.append(file)   

print('ENWIKI_ART_FNMS:', len(ENWIKI_ART_FNMS))
print('ENWIKI_RED_FNMS:', len(ENWIKI_RED_FNMS))

df_en_id_name = pd.read_csv(PATH_TO_DATA_EN+"enwiki-" + data_time + "-id_name.csv", encoding='utf-8', sep="^")
df_en_id_name.head()

df_en_id_name.shape

# go to alias grabber file

df_en_id_orig_title_alias = pd.read_csv(PATH_TO_DATA_EN+"enwiki-" + data_time + "-id_orig_title_alias.csv")
df_en_id_orig_title_alias.head()

df_en_id_orig_title_alias.columns = ['id', 'title']
df_en_id_name_all = df_en_id_name.append(df_en_id_orig_title_alias)
df_en_id_name_all.shape

#ignored redirections
#df_en_id_name_all = df_en_id_name

df_en_id_name_all['title'] = df_en_id_name_all['title'].str.strip()

df_en_id_name_all['title_lc'] = df_en_id_name_all['title'].str.lower()
df_en_id_name_all = df_en_id_name_all[['id', 'title_lc']]
df_en_id_name_all = df_en_id_name_all.drop_duplicates()
df_en_id_name_all.shape

df_en_id_name_all.head()

for fn in ENWIKI_ART_FNMS:
    fn = PATH_TO_DATA_EN+fn
    print(fn)
    fn_new = unpack(fn)
    df_articles = pd.read_csv(fn_new, encoding='ISO-8859-1', quotechar="\"")
    df_articles['link_val'] = df_articles['link_val'].str.split("#").str[0]
    #df_articles = df_articles[df_articles['link_val'].notna()]
    df_articles = df_articles[df_articles['link_val'].str.strip() != '']
    df_articles['link_pos_perc'] = (df_articles['link_pos']/df_articles['text_len']).round(3)       
    df_articles['link_val_lc'] = df_articles['link_val'].str.lower()
    df_en_titles_series = df_articles.iloc[:,7]
    df_en_titles_series_stripped = df_en_titles_series.str.replace('  ', ' ')
    df_titles_stripped = df_en_titles_series_stripped.to_frame()
    df_articles_less = df_articles[['id', 'title', 'link_val']]
    df_articles = pd.concat([df_articles_less, df_titles_stripped], axis=1)
    df_articles['link_val_lc'] = df_articles['link_val_lc'].str.strip()
    df_articles = pd.merge(df_articles, df_en_id_name_all, how='left', left_on=['link_val_lc'], right_on=['title_lc'])
    df_articles = df_articles[['id_x', 'id_y', 'link_val']]
    df_articles.columns = ['id', 'link_id', 'link_val']
    df_articles = df_articles[pd.notnull(df_articles['link_val'])]
    df_articles[['link_id']] = df_articles[['link_id']].fillna(-1).astype(int)
    df_articles['is_red_link'] = np.where(df_articles['link_id'] == -1, True, False)
    df_articles.to_csv(PATH_TO_DATA_EN+fn_new, index = False, sep='^')
    pack_and_remove(fn_new)

!pwd

#??????????????????

df_id_name = pd.read_csv(PATH_TO_DATA+'enwiki-' + data_time + '-id_name.csv', encoding='ISO-8859-1', quotechar="'", escapechar ="\\")

df_id_name.head()

fn = PATH_TO_DATA_EN+"enwiki-" + data_time + "-langlinks_en.csv"
df_en_langlinks = pd.read_csv(fn, encoding='UTF-16', sep="^")

df_en_langlinks = df_en_langlinks[pd.notnull(df_en_langlinks['ll_title'])]

df_en_langlinks.head()

df_en_en_links = pd.merge(df_en_langlinks, df_id_name, how='left', left_on=['ll_title'], right_on=['title'])

df_en_en_links = df_en_en_links[['ll_from', 'id']]
df_en_en_links.columns = ['id_en', 'id_en']
df_en_en_links[['id_en']] = df_en_en_links[['id_en']].fillna(-1).astype(int)
df_en_en_links[['id_en']] = df_en_en_links[['id_en']].fillna(-1).astype(int)
df_en_en_links.to_csv(PATH_TO_DATA_EN + data_time + "-langlinks_en_en.csv", index = False, sep'^')
pack_and_remove(fn_new)

df_en_langlinks.to_csv(PATH_TO_DATA_EN + data_time + "-langlinks_en_en.csv", index = False, sep'^')
