==> notebooks/data retrieval/red link-candidate pairs/07. alias-grabber-en.ipynb <==
import codecs
import csv
import time
import os
import re
import gzip
import pandas as pd
import numpy as np

PATH_TO_DATA = '/media/andrii/earth/Katia/CS_MasterThesis/data/enwiki/en_parsed/'
data_time = '20180920'

ENWIKI_ALI_FNMS = []
for file in os.listdir(PATH_TO_DATA):
    if file.endswith("red.csv.gz"):
        ENWIKI_ALI_FNMS.append(file)  

len(ENWIKI_ALI_FNMS)

EN_ID_NAME = "enwiki-" + data_time + "-id_name.csv"  

def unpack(file_name):
    file_name_new = file_name.replace(".gz","")
    with gzip.open(file_name, 'rb') as f_in, open(file_name_new, 'wb') as f_out:
        f_out.writelines(f_in)
    return file_name_new

def pack_and_remove(file_name):
    file_name_new = file_name+'.gz'
    with open(file_name, 'rb') as f_in, gzip.open(file_name_new, 'wb') as f_out:
        f_out.writelines(f_in)
    os.remove(file_name)
    return file_name_new

df_en_id_name = pd.read_csv(PATH_TO_DATA+EN_ID_NAME, encoding='utf-8', sep="^")
df_en_id_name.head()

df_id_alias_title_alias = None
df_id_alias_id_orig = None
df_id_orig_title_alias = None
for fn in ENWIKI_ALI_FNMS:
    fn = PATH_TO_DATA+fn
    print(fn)
    fn_new = unpack(fn)
    df_aliases = pd.read_csv(fn_new, encoding='UTF-8', quotechar="\"")
    df_aliases = pd.merge(df_aliases, df_en_id_name, how='left', left_on=['redirect'], right_on=['title'])    
    df_aliases = df_aliases[['id_x', 'id_y', 'title_x', 'redirect']]
    df_aliases.columns = ['id_alias', 'id_orig', 'title_alias', 'title_orig']    
    
    df_id_alias_title_alias_tmp = df_aliases[['id_alias', 'title_alias']]
    if df_id_alias_title_alias is not None:
        df_id_alias_title_alias = df_id_alias_title_alias.append(df_id_alias_title_alias_tmp)        
    else:
        df_id_alias_title_alias = df_id_alias_title_alias_tmp
    print("df_id_alias_title_alias size: {}".format(df_id_alias_title_alias.shape))
    
    df_id_alias_id_orig_tmp = df_aliases[['id_alias', 'id_orig']]
    if df_id_alias_id_orig is not None:
        df_id_alias_id_orig = df_id_alias_id_orig.append(df_id_alias_id_orig_tmp)        
    else:
        df_id_alias_id_orig = df_id_alias_id_orig_tmp
    print("df_id_alias_id_orig size: {}".format(df_id_alias_id_orig.shape))
    
    df_id_orig_title_alias_tmp = df_aliases[['id_orig', 'title_alias']]
    if df_id_orig_title_alias is not None:
        df_id_orig_title_alias = df_id_orig_title_alias.append(df_id_orig_title_alias_tmp)        
    else:
        df_id_orig_title_alias = df_id_orig_title_alias_tmp
    print("df_id_orig_title_alias size: {}".format(df_id_orig_title_alias.shape))
    os.remove(fn_new)

df_id_alias_title_alias.head()

df_id_alias_title_alias.to_csv(PATH_TO_DATA+'enwiki-' + data_time + '-id_alias_title_alias.csv', index = False, encoding='UTF-8', quotechar="\"")

df_id_alias_id_orig[['id_orig']] = df_id_alias_id_orig[['id_orig']].fillna(-1).astype(int)

df_id_alias_id_orig.head()

df_id_alias_id_orig.to_csv(PATH_TO_DATA+'enwiki-' + data_time + '-id_alias_id_orig.csv', index = False, encoding='UTF-8', quotechar="\"")

df_id_orig_title_alias[['id_orig']] = df_id_orig_title_alias[['id_orig']].fillna(-1).astype(int)

df_id_orig_title_alias.shape

df_id_orig_title_alias = df_id_orig_title_alias.drop_duplicates()
df_id_orig_title_alias.shape

df_id_orig_title_alias.head()

df_id_orig_title_alias.to_csv(PATH_TO_DATA+'enwiki-' + data_time + '-id_orig_title_alias.csv', index = False, encoding='UTF-8', quotechar="\"")
