==> notebooks/data retrieval/red link-candidate pairs/08. ukredlink_JaccardModel.ipynb <==
import pandas as pd
import gzip
import numpy as np
import time
import os
import re
import progressbar

%load_ext autoreload
%autoreload 2

import sys
sys.path.insert(0, '/media/andrii/earth/Katia/CS_MasterThesis/Red_links_Project_for_Wiki_draft/py')
import utils

### Load en_wiki

#path to data
PATH_TO_DATA = '/media/andrii/earth/Katia/CS_MasterThesis/data/'
PATH_TO_DATA_UK = PATH_TO_DATA+"ukwiki/"
PATH_TO_DATA_EN = PATH_TO_DATA+"enwiki/en_parsed/"

data_time = '20180920'

# get list of archives to unpack for uk wiki
ENWIKI_ART_FNMS = []
for file in os.listdir(PATH_TO_DATA_EN):
    if re.match(r"enwiki-20180920-pages-meta-current\d+\.xml-p\d+p\d+\.bz2_art\.csv.gz", file):
        ENWIKI_ART_FNMS.append(file) 
 

len(ENWIKI_ART_FNMS)

EN_ID_NAME = "enwiki-"+data_time+"-id_name.csv"

def unpack(file_name):
    file_name_new = file_name.replace(".gz","")
    with gzip.open(file_name, 'rb') as f_in, open(file_name_new, 'wb') as f_out:
        f_out.writelines(f_in)
    return file_name_new

start_time = time.time()
# loading all articles to dataframe
df_en = pd.DataFrame()
n_files = len(ENWIKI_ART_FNMS)
print(n_files)

for index in range(n_files):
    print(str(index+1) + '/' + str(n_files))
    fn = ENWIKI_ART_FNMS[index]
    fn = PATH_TO_DATA_EN+fn
    print(fn)
    fn_new = unpack(fn)
    df_articles = pd.read_csv(fn_new, encoding='UTF-8', sep='^')  
    df_articles = df_articles[df_articles['is_red_link']==False][['id','link_id']]
    df_en = pd.concat((df_en, df_articles))
    print("df_articles size: {}. df_articles columns: {}\n".format(df_articles.shape, list(df_articles.columns)))
    os.remove(fn_new)
    
print('Total time: %.1f minutes' % ((time.time() - start_time)/60))

print(df_en.shape)
print(df_en.columns)

# article names
df_en_name = pd.read_csv(PATH_TO_DATA_EN+EN_ID_NAME,  encoding='UTF-8', quotechar="\"", sep='^')
df_en_name = df_en_name[['id','title']]
print(df_en_name.shape)
print(df_en_name.columns)

### Load uk_wiki

UKWIKI_ART_FNMS = []
for file in os.listdir(PATH_TO_DATA_UK):
    if re.match(r"ukwiki-20180920-pages-links.csv", file):
        UKWIKI_ART_FNMS.append(file)

UKWIKI_ART_FNMS

UK_ID_NAME = "ukwiki-"+data_time+"-id_name.csv"
UK_ID_NAME

start_time = time.time()

# loading existing articles and red links to dataframes
df_uk_blue = pd.DataFrame()
df_uk_red = pd.DataFrame()
n_files = len(UKWIKI_ART_FNMS)
for index in range(n_files):
    print(str(index+1) + '/' + str(n_files))
    fn = UKWIKI_ART_FNMS[index]
    fn = PATH_TO_DATA_UK+fn
    print(fn)
    df_articles = pd.read_csv(fn)
    
    df_articles_blue = df_articles[df_articles['is_red_link']==False]
    df_uk_blue = pd.concat((df_uk_blue, df_articles_blue[['id','link_id']]))
    df_articles_red = df_articles[df_articles['is_red_link']]
    df_uk_red = pd.concat((df_uk_red, df_articles_red[['id','link_val']]))
    
    print("df_articles size: {}. df_articles columns: {}\n".format(df_articles.shape, list(df_articles.columns)))

print('Total time: %.1f minutes' % ((time.time() - start_time)/60))

print(df_uk_blue.shape)
print(df_uk_blue.columns)

print(df_uk_red.shape)
print(df_uk_red.columns)

# article names
df_uk_name = pd.read_csv(PATH_TO_DATA_UK+UK_ID_NAME, sep='^')

print(df_uk_name.shape)
print(df_uk_name.columns)

# # add ids for red links

# # find maximum id among blue links
max_id = np.max(df_uk_name['id'])

# # find unique titles among red links, create df with
# red_links = df_uk_red['link_val'].unique()
# red_links_ids = pd.DataFrame({'title': red_links, 
#                               'id': np.arange(max_id+1, len(red_links)+max_id+1)})
# red_links_ids['is_red_link'] = True

# # create dataframe with titles and red link tickets / add it to article titles df
# df_uk_name['is_red_link'] = False
# df_uk_name = pd.concat((df_uk_name,red_links_ids))
# df_uk_name = df_uk_name.reset_index(drop=True)

# print(df_uk_name.shape)
# print(df_uk_name.columns)


# # add link ids to df_uk_red
# df_uk_red = df_uk_red.merge(right = red_links_ids, left_on = 'link_val', right_on = 'title', how = 'left')
# df_uk_red = df_uk_red[['id_x','id_y']]
# df_uk_red.columns = ['id', 'link_id']

df_uk_red = pd.read_csv(PATH_TO_DATA_UK+'all_uk_redlinks_with_parent_ids.csv')

df_uk_red.head()

df_uk_red = df_uk_red[['id_parent', 'red_link_id']]
df_uk_red.columns = ['id', 'link_id']

# concatenate red and blue article to one df
df_uk = pd.concat((df_uk_red, df_uk_blue))

print(df_uk.shape)
print(df_uk_blue.shape[0] + df_uk_red.shape[0])

del df_uk_red, df_uk_blue

### Load uk-en correpondences

fn = PATH_TO_DATA_UK+'20180920-langlinks_uk_en.csv'
df_link = pd.read_csv(fn, encoding='UTF-8', quotechar='\'')
print(df_link.shape)
print(df_link.columns)

df_link.head()

# # remove values links to negative values
# df_link = df_link[(df_link['id_en']>0) & (df_link['id_en']!=49244)]
# print(df_link.shape)

df_en_name.head()

# add uk correspondences to en article names
df_en_name = pd.merge(df_en_name, df_link, left_on = 'title', right_on = 'll_title', how = 'left')

df_en_name.head()

df_en_name_ = df_en_name[['id','title', 'll_from']]
df_en_name_.columns = ['id', 'title', 'id_uk']

df_en_name_.head()

# several en articles have multiple corresponding uk acticles, just remove oldest one uk article
df_en_translated = (df_en_name_[~df_en_name_['id_uk'].isnull()])[['id', 'id_uk']]
df_en_translated.columns = ['id_en','id_uk']
df_en_translated = df_en_translated.sort_values(by = 'id_en')
df_en_translated = df_en_translated.drop_duplicates(keep = 'first', subset = ['id_uk'])

# add en correspondences to uk article names

df_articles.head()

df_uk_name = df_articles.merge(right = df_en_translated, left_on = 'id', right_on = 'id_uk', how = 'left')

df_uk_name.head()

df_uk_name = df_uk_name[['id','link_val','is_red_link','id_en']]
df_uk_name.columns = ['id','title','is_red_link','id_en']
df_uk_name = df_uk_name.sort_values(by = 'id')

df_uk_name.head()

df_en_name_.head()

# save names to files
df_uk_name.to_csv(PATH_TO_DATA_UK+'uk_names.csv', header=True, index=False, sep='^')

df_en_name_.to_csv(PATH_TO_DATA_EN+'en_names.csv', header=True, index=False, sep='^')

df_uk_name.head()

# load names 
df_uk_name = pd.read_csv(PATH_TO_DATA_UK+'uk_names.csv', sep='^')
df_en_name = pd.read_csv(PATH_TO_DATA_EN+'en_names.csv', sep='^')

### Encode en nontranslated article by its incoming links

# select non-translated articles
en_nontranslated = np.array(df_en_name[df_en_name['id_uk'].isnull()]['id'])
print('Nontranslated en acticles: %6d' % (len(en_nontranslated)))

# select non-translated acticles that have at least 20 distinct incoming links
en_nontranslated = df_en[df_en['link_id'].isin(en_nontranslated)].groupby('link_id') \
            .agg({'id': lambda x: x.nunique()})
en_nontranslated = en_nontranslated.reset_index()
en_nontranslated.columns = ['id','n_incoming']
en_nontranslated = en_nontranslated[en_nontranslated['n_incoming']>=20]
en_nontranslated = np.array(en_nontranslated['id'])
en_nontranslated_ids = np.sort(en_nontranslated)
print('Nontranslated en acticles with at least 20 incoming links: %6d' % (len(en_nontranslated)))

# encode each nontranslated article by its incoming links
en_nontranslated_encoding_df = df_en[df_en['link_id'].isin(en_nontranslated)]
en_nontranslated_encoding_df = en_nontranslated_encoding_df.sort_values(by = ['link_id','id'])
en_nontranslated_encoding_df = en_nontranslated_encoding_df.reset_index(drop = True)
indices = np.array(en_nontranslated_encoding_df.drop_duplicates(keep='first', subset=['link_id']).index)

en_nontranslated_encoding = []
pbar = progressbar.ProgressBar()
for i in pbar(range(0, len(indices))):
    if i == len(indices) - 1:
        this_encoding = set(en_nontranslated_encoding_df.iloc[indices[i]:]['id'])
    else: 
        this_encoding = set(en_nontranslated_encoding_df.iloc[indices[i]:indices[i+1]]['id'])
    en_nontranslated_encoding.append(this_encoding)
    
del en_nontranslated_encoding_df, en_nontranslated

# save encoding to file
np.save(PATH_TO_DATA_EN+'en_nontranslated_encoding.npy', arr=en_nontranslated_encoding)
np.save(PATH_TO_DATA_EN+'en_nontranslated_ids.npy', arr=en_nontranslated_ids)

# load encoding from file
en_nontranslated_encoding = np.load(PATH_TO_DATA_EN+'en_nontranslated_encoding.npy', allow_pickle=True)
en_nontranslated_ids = np.load(PATH_TO_DATA_EN+'en_nontranslated_ids.npy', allow_pickle=True)

### Encode red links incoming en links

# select only red links from uk_wiki
df_uk_red = df_uk[df_uk['link_id']>max_id]
uk_red_ids = df_uk_red['link_id'].unique()
print('Number of red links in uk wiki: %d' %(len(uk_red_ids)))

df_uk_red = df_uk_red.drop_duplicates()

len(df_uk_red)

df_uk_red.head()

df_uk_name.head()

ukredlinks_with_enids_sample = pd.read_csv(PATH_TO_DATA_EN+'ukredlinks_with_enids_sample.csv', sep='^')

ukredlinks_with_enids_sample.head()

# for every red link from the sample find its correponding en incoming link (only distinct)
df_uk_red_encoding = ukredlinks_with_enids_sample[['id', 'red_link_id']]
df_uk_red_encoding.columns = ['id_en','link_id']
df_uk_red_encoding = df_uk_red_encoding[~df_uk_red_encoding['id_en'].isnull()]
df_uk_red_encoding = df_uk_red_encoding.drop_duplicates()

df_uk_red_encoding.shape

df_uk_red_encoding.head()

# for every red link from the sample calculate number of incoming links
df_uk_red_incoming = df_uk_red_encoding['link_id'].value_counts().reset_index()
df_uk_red_incoming.columns = ['id','n_incoming']

# select only those red links that can be encoded with at least 20 en incoming links
uk_red_ids = np.array(df_uk_red_incoming[df_uk_red_incoming['n_incoming']>=20]['id'])
print('Number of red links with at least 20 distinct incoming en links: %d' % (len(uk_red_ids)))

# encode top red links using incoming en links
uk_red_ids = np.sort(uk_red_ids)
df_uk_red_encoding = df_uk_red_encoding[df_uk_red_encoding['link_id'].isin(uk_red_ids)]
df_uk_red_encoding = df_uk_red_encoding.sort_values(by = 'link_id').reset_index(drop = True)
indices = np.array(df_uk_red_encoding.drop_duplicates(keep='first', subset=['link_id']).index)

uk_red_encoding = []
pbar = progressbar.ProgressBar()
for i in pbar(range(0, len(indices))):
    if i == len(indices)-1:
        this_encoding = set(df_uk_red_encoding.iloc[indices[i]:]['id_en'])
    else:
        this_encoding = set(df_uk_red_encoding.iloc[indices[i]:indices[i+1]]['id_en'])
    uk_red_encoding.append(this_encoding)

del df_uk_red_incoming, df_uk_red_encoding

# save encoding to file
np.save(file='uk_red_ids.npy', arr=uk_red_ids)
np.save(file='uk_red_encoding.npy', arr=uk_red_encoding)

# load encoding from file
uk_red_ids = np.load('uk_red_ids.npy', allow_pickle=True)
uk_red_encoding = np.load('uk_red_encoding.npy', allow_pickle=True)

df_uk_name.head()

print(len(list(df_uk_name[df_uk_name['id'].isin(uk_red_ids)]['title'])))

red_links_names = ukredlinks_with_enids_sample[ukredlinks_with_enids_sample['red_link_id'].isin(uk_red_ids)]['red_link_name'].to_frame()
red_links_names = red_links_names.drop_duplicates()
list_of_red_link = red_links_names['red_link_name'].tolist()

# create df with incoming uk ids and names for every red link
incoming_uk_link_names = []
n_incoming_uk = []
incoming_en_link_names = []
n_incoming_en = []

pbar = progressbar.ProgressBar()
for i in pbar(range(len(uk_red_ids))):
    red_id = uk_red_ids[i]
    incoming_ids = list(df_uk_red[df_uk_red['link_id'] == red_id]['id'])
    incoming_names = set(df_uk_name[df_uk_name['id'].isin(incoming_ids)]['title'])
    incoming_en_names = set(df_en_name[df_en_name['id'].isin(uk_red_encoding[i])]['title'])
              
    incoming_uk_link_names.append(incoming_names)
    n_incoming_uk.append(len(incoming_names))
    incoming_en_link_names.append(incoming_en_names) 
    n_incoming_en.append(len(incoming_en_names))

df_red_links = pd.DataFrame({'red_link_id': uk_red_ids, 
                          'red_link_name': list_of_red_link,
                          'incoming_uk_names': incoming_uk_link_names,
                          'n_incoming_uk': n_incoming_uk,
                          'incoming_en_names': incoming_en_link_names,
                          'n_incoming_en': n_incoming_en
                          })
              
df_red_links = df_red_links[['red_link_id','red_link_name','incoming_uk_names',
                             'n_incoming_uk', 'incoming_en_names', 'n_incoming_en']]

print(df_red_links.shape)
df_red_links.head(100)

# save red links incoming links to files
df_red_links.to_csv(PATH_TO_DATA_UK+'red_links_summary_jaccard_model.csv.gz', compression='gzip', header=True, index=False)

# load 
df_red_links = pd.read_csv(PATH_TO_DATA_UK+'red_links_summary_jaccard_model.csv.gz')

### Find most similar uk articles for red links

# jaccard similarity
def jaccard(a, b):
    c = a.intersection(b)
    return len(c) / (len(a) + len(b) - len(c))

! Version 1

en_article_list = []
en_article_found = []

pbar = progressbar.ProgressBar()
for j in pbar(range(len(uk_red_ids))):
    red_link = uk_red_ids[j]
    red_link_encoding = uk_red_encoding[j]
    similarities = []
    for i in range(len(en_nontranslated_ids)):
        similarity = jaccard(red_link_encoding, en_nontranslated_encoding[i])
        similarities.append(similarity)
    
    # select top 5 most similar uk article, zero similarities excluded
    similarities = np.array(similarities)
    indices = np.argsort(-similarities)[:5]
    score = np.round(similarities[indices],3)
    n_nonzero = np.sum(score>0)
    indices = np.sort(indices[:n_nonzero])

    # scores and article names sorted by article ids
    score = np.round(similarities[indices],3)
    en_article_ids = en_nontranslated_ids[indices]
    en_article_names = list(df_en_name[df_en_name['id'].isin(en_article_ids)]['title'])
    
    # sort by jaccard similarity
    d = dict(zip(en_article_names, score))
    d_sorted = sorted(((value, key) for (key,value) in d.items()), reverse=True)
    
    # save to list
    en_article_list.append(d_sorted)
    
    if len(en_article_names)>0:
        en_article_found.append(True)
    else:
        en_article_found.append(False)

        
# create df to show results
df_similarities = pd.DataFrame({'red_link_id': uk_red_ids,
                   'en_similar_found': en_article_found, 
                   'en_similar': en_article_list
                  })
    
    

# version for getting all candidates


en_article_list = []
en_article_found = []

pbar = progressbar.ProgressBar()
for j in pbar(range(len(uk_red_ids))):
    red_link = uk_red_ids[j]
    red_link_encoding = uk_red_encoding[j]
    similarities = []
    for i in range(len(en_nontranslated_ids)):
        similarity = jaccard(red_link_encoding, en_nontranslated_encoding[i])
        similarities.append(similarity)
    
    # zero similarities excluded
    similarities = np.array(similarities)
    indices = np.argsort(-similarities)
    score = np.round(similarities[indices],3)
    n_nonzero = np.sum(score>0)
    indices = np.sort(indices[:n_nonzero])

    # scores and article names sorted by article ids
    score = np.round(similarities[indices],3)
    en_article_ids = en_nontranslated_ids[indices]
    en_article_names = list(df_en_name[df_en_name['id'].isin(en_article_ids)]['title'])
    
    # sort by jaccard similarity
    d = dict(zip(en_article_names, score))
    d_sorted = sorted(((value, key) for (key,value) in d.items()), reverse=True)
    
    # save to list
    en_article_list.append(d_sorted)
    
    if len(en_article_names)>0:
        en_article_found.append(True)
    else:
        en_article_found.append(False)

        
# create df to show results
df_similarities = pd.DataFrame({'red_link_id': uk_red_ids,
                   'en_similar_found': en_article_found, 
                   'en_similar': en_article_list
                  })
    
    

df_red_results = df_red_links.merge(right=df_similarities, right_on = 'red_link_id',
                                     left_on = 'red_link_id', how = 'left')

df_red_results.head()

df_red_results['en_similar_found'].value_counts()

# save results with all candidates to file
df_red_results.to_csv(PATH_TO_DATA_UK+'uk_red_links_results_all_candidates.csv', header=True, index=False, encoding = 'UTF-8', sep='^')

# save results to file
df_red_results.to_csv(PATH_TO_DATA_UK+'uk_red_links_results.csv', header=True, index=False, encoding = 'UTF-8', sep='^')

# load 
df_red_results = pd.read_csv(PATH_TO_DATA_UK+'uk_red_links_results.csv', encoding = 'UTF-8', sep='^')

df_red_results = df_red_results[['red_link_name', 'en_similar']]

# save results to file
df_red_results.to_csv(PATH_TO_DATA_UK+'uk_red_links_results_lighter.csv', header=True, index=False, encoding = 'UTF-8', sep='^')
