==> notebooks/data retrieval/red link-candidate pairs/09. get_candidate_pairs.ipynb <==
import pandas as pd
import numpy as np

import sys
sys.path.insert(0, '/media/andrii/earth/Katia/CS_MasterThesis/Red_links_Project_for_Wiki_draft/py')
import utils

PATH_TO_DATA = '/media/andrii/earth/Katia/CS_MasterThesis/data/ukwiki/'

df_red_candidates = pd.read_csv(PATH_TO_DATA+'uk_red_links_results_all_candidates.csv', encoding = 'UTF-8', sep='^')

df_red_candidates.head()

df_red_candidates

df_red_candidates['en_similar'][94]

all_results = {}
all_results_list = [[]] * df_red_candidates.shape[0]
for i in range(32, df_red_candidates.shape[0]):
    print(i)
    if df_red_candidates['en_similar'][i] == '[]':
        all_results[df_red_candidates['red_link_name'][i]] = []
        all_results_list[i] = []
    else:
        res = utils.extract_list(df_red_candidates['en_similar'][i])
        all_results[df_red_candidates['red_link_name'][i]] = res
        all_results_list[i] = res

rl_with_cand = pd.DataFrame.from_dict(all_results, orient='index')

rl_with_cand.head()

num_df_list = utils.divide_number(rl_with_cand.shape[0], 100)
i = len(num_df_list)
start_index = 0
num_rows_proceed = num_df_list[i-1]

while i > 0:
    df_all = pd.DataFrame(columns=['red_link', 'candidate'])
    print('start_index =', start_index)
    print('num_rows_proceed =', num_rows_proceed)
    for row in range(start_index, num_rows_proceed):
        num_rows = rl_with_cand.count(axis=1)[row]

        df = pd.DataFrame(np.nan, index=range(0,num_rows), columns=['red_link', 'candidate'])

        red_link = rl_with_cand.index[row]
        # fill nans with data
        for column in range(num_rows):
            candidate = rl_with_cand.iloc[row,column]

            df.iloc[column,0] = red_link
            df.iloc[column,1] = candidate

        df_all = pd.concat([df_all, df])
        
    df_all.to_csv(PATH_TO_DATA+'candidates/'+'candidates_'+str(i)+'.csv', index=False, sep='^')
    i -= 1
    start_index = num_rows_proceed
    num_rows_proceed = num_rows_proceed + num_df_list[i-1]

# results is files with red link-candidate pairs 'candidates_'+(1-101)+'.csv'
