==> notebooks/data retrieval/red link-candidate pairs/10. add_ground_truth+train_test_split.ipynb <==
## Split data into train and test sets with ground truth

import pandas as pd
import os

import sys
sys.path.insert(0, '/media/andrii/earth/Katia/CS_MasterThesis/Red_links_Project_for_Wiki_draft/py')
import utils

PATH_TO_DATA = '/media/andrii/earth/Katia/CS_MasterThesis/data/ukwiki/'

red_links_with_gt = pd.read_csv(PATH_TO_DATA+'ukredlinks_with_ground_truth.csv', sep='^')

red_links_with_gt.columns = ['red link name', 'ground truth']

rl_with_candidates = os.listdir(PATH_TO_DATA+'candidates/')

df_with_all_candidates = utils.load_csvs_to_df_general(PATH_TO_DATA+'candidates/', rl_with_candidates)

df_with_all_candidates.head()

df_with_all_candidates.shape

df_with_all_candidates.shape

rl_all_pairs_gt = pd.merge(df_with_all_candidates, red_links_with_gt, how='left', left_on=['red_link'], right_on=['red link name'])

rl_all_pairs_gt = rl_all_pairs_gt[['red_link', 'candidate', 'ground truth']]

# add true label to train pairs
gt_results = []
for index, row in rl_all_pairs_gt.iterrows():
    gt_result = row['candidate'] == row['ground truth']
    gt_results.append(gt_result)

gt_df = pd.DataFrame(gt_results)
gt_df.columns = ['true_label']

with_gt_df = pd.concat([rl_all_pairs_gt, gt_df], axis=1)

with_gt_df.head()

utils.train_test_split(PATH_TO_DATA, with_gt_df)

with_gt_df.to_csv(PATH_TO_DATA+'ukredlinks_with_all_candidates&true_label', index=False, sep='^')
